+++ src/vi/k2600.ttm/legacy.mstest/Tests/Asserts_driver.cs
@@ -637,8 +637,8 @@ internal static partial class Asserts
             if ( outcomeValue == ( int ) FirmwareOutcomes.Okay )
             {
                 bitValue = ( int ) cc.isr.VI.BufferElementStatusBits.OverTemp;
-                Assert.AreEqual( bitValue & statusValue, 0,
-                    $"status {statusValue} must not have the over temp bit {bitValue} set if measurement was made." );
+                Assert.AreEqual( 0, bitValue & statusValue,
+                    $"status {statusValue} must not have the over-temp bit {bitValue} set if measurement was made." );
                 // compliance is acceptable
                 // bitValue = 64;
                 // Assert.IsTrue( 0 == (bitValue & statusValue), $"status {statusValue} must not have the compliance bit {bitValue} set if measurement was made." );