+++ changelog.md
@@ -2,8 +2,12 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9279]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9284]: https://www.github.com/atecoder/dn.vi.ivi
 
+## [8.1.9284 - 2025-06-02
+- Access Subsystem Base:
+  - Add registration and certification methods.
+  
 ## [8.1.9279] - 2025-05-28
 - VI.Device Access Subsystem Base:
   - Consolidate the registration code.